new uvs2 releases : uppsrc-2598  tutorial-38  examples-141  reference-113
git-svn-id: svn://ultimatepp.org/upp/trunk@300 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Core/Algo.h
@@ -897,7 +897,7 @@ struct StableSortLess_ {
 template <class I, class Less, class T>
 void StableSort_(I begin, I end, const Less& less, const T *)
 {
-	int count = end - begin;
+	int count = (int)(uintptr_t)(end - begin);
 	Buffer<int> h(count);
 	for(int i = 0; i < count; i++)
 		h[i] = i;
